security and complexity updates
