+++ routes/access_api.py
@@ -1,19 +1,14 @@
-import asyncio
-import contextlib
-import uuid
 from urllib.parse import quote_plus
 
-from cassandra.query import SimpleStatement
 from fastapi import APIRouter, HTTPException, BackgroundTasks
 from pydantic import BaseModel, Field
 from sqlalchemy import create_engine
 from sqlalchemy.exc import OperationalError
 
-from database import database
-from database.database import execute_query, _session
 from models.models import AccessCredentials, SiteConfig
 from serializers.access_credentials_serializer import access_credential_list_entity, systems_list_entity, system_entity, \
     access_credential_entity
+from utils.csv_upload_handler import upload_data
 
 router = APIRouter()
 
@@ -72,32 +67,6 @@ async def add_connection(data: SaveDBConnection):
 class SaveCSVUploadData(BaseModel):
     name: str = Field(..., description="Name of connection")
     data: list = Field(..., description="")
-
-
-def create_table(data):
-    columns = ', '.join([f'{key} TEXT' for key in data.data[0].keys()])
-    _session.execute(f"""
-    CREATE TABLE IF NOT EXISTS {data.name}_CSV_EXTRACT (
-        generated_id_unique UUID PRIMARY KEY, -- primary key generated for this table
-        {columns}
-    )
-    """)
-
-
-def upload_data(data):
-    try:
-        create_table(data)
-        for record in data.data:
-            columns = ', '.join(record.keys())
-            placeholders = ', '.join(['%s'] * len(record))
-            query = SimpleStatement(f"""
-                INSERT INTO {data.name}_CSV_EXTRACT (generated_id_unique, {columns})
-                VALUES (%s, {placeholders})
-            """)
-            values = (uuid.uuid4(), *record.values())
-            _session.execute(query, values)
-    except Exception as e:
-        print(e)
 
 
 @router.post('/upload_csv')
